스킬애니메이션 처리 수정 및 TransparentBlt RGB값 변환 수정 중

+++ Maplestory WinAPI Project/Animation.h
@@ -35,9 +35,9 @@ public:
 
 	bool IsComplete() { return mbComplete; }
 	void SetAnimator(Animator* animator) { mAnimator = animator; }
-	void SetR(float r) { float R = r; }
-	void SetG(float g) { float G = g; }
-	void SetB(float b) { float B = b; }
+	void SetR(int r) { int R = r; }
+	void SetG(int g) { int G = g; }
+	void SetB(int b) { int B = b; }
 
 private:
 	Animator* mAnimator;
@@ -46,8 +46,8 @@ private:
 	float mTime;
 	bool mbComplete;
 	int mSpriteIndex;
-	float R;
-	float G;
-	float B;
+	int R;
+	int G;
+	int B;
 	};
